+++ app/components/tasks.hbs
@@ -5,6 +5,7 @@
       <div>
         <Task::Holder
           @task={{task}}
+          @onTaskChange={{@onTaskChange}}
           @onStausChange={{this.onStatusChange}}
           @onTaskUpdate={{@onTaskUpdate}}
           @isLoading={{@isLoading}}
