+++ app/templates/tasks.hbs
@@ -10,7 +10,8 @@
   <Tasks
     @userSelectedTask={{this.userSelectedTask}}
     @tasksToShow={{this.tasksToShow}}
-    @updateLoader={{this.updateLoader}}
+    @onTaskChange={{this.onTaskChange}}
+    @onTaskUpdate={{this.handleUpdateTask}}
     @isLoading={{this.isLoading}}
     @disabled={{this.disabled}}
   />
